Break up update checker and move System health checks to a package closer to the update checks
