+++ game-core/src/main/java/games/strategy/engine/auto/update/SystemHealthCheck.java
@@ -1,0 +1,4 @@
+package games.strategy.engine.auto.update;
+
+public class SystemHealthCheck {
+}
