+++ game-core/src/main/java/games/strategy/engine/auto/update/UpdateChecks.java
@@ -2,22 +2,11 @@ package games.strategy.engine.auto.update;
 
 import static games.strategy.engine.framework.ArgParser.CliProperties.TRIPLEA_GAME;
 
-import java.time.LocalDateTime;
-import java.time.temporal.ChronoField;
 import java.util.Collection;
 
-import javax.swing.JOptionPane;
-import javax.swing.SwingUtilities;
-
 import games.strategy.debug.ClientLogger;
-import games.strategy.engine.ClientContext;
-import games.strategy.engine.framework.EngineVersionProperties;
-import games.strategy.engine.framework.map.download.DownloadMapsWindow;
+import games.strategy.engine.auto.health.check.LocalSystemChecker;
 import games.strategy.engine.framework.map.download.MapDownloadController;
-import games.strategy.engine.framework.systemcheck.LocalSystemChecker;
-import games.strategy.triplea.settings.ClientSetting;
-import games.strategy.ui.SwingComponents;
-import games.strategy.util.EventThreadJOptionPane;
 
 /**
  * Runs background update checks and would prompt user if anything needs to be updated.
@@ -52,63 +41,11 @@ public class UpdateChecks {
       return;
     }
 
-    checkForTutorialMap();
-    checkForLatestEngineVersionOut();
+    TutorialMapDecision.checkForTutorialMap();
+    EngineVersionCheck.checkForLatestEngineVersionOut();
 
     if (UpdateCheckDecision.shouldRunMapUpdateCheck()) {
       MapDownloadController.checkDownloadedMapsAreLatest();
     }
-  }
-
-  /**
-   * Returns true if we are out of date or this is the first time this triplea has ever been run.
-   */
-  private static void checkForLatestEngineVersionOut() {
-    try {
-      final boolean firstTimeThisVersion = ClientSetting.TRIPLEA_FIRST_TIME_THIS_VERSION_PROPERTY.booleanValue();
-      // check at most once per 2 days (but still allow a 'first run message' for a new version of triplea)
-      final LocalDateTime localDateTime = LocalDateTime.now();
-      final int year = localDateTime.get(ChronoField.YEAR);
-      final int day = localDateTime.get(ChronoField.DAY_OF_YEAR);
-      // format year:day
-      final String lastCheckTime = ClientSetting.TRIPLEA_LAST_CHECK_FOR_ENGINE_UPDATE.value();
-      if (!firstTimeThisVersion && lastCheckTime.trim().length() > 0) {
-        final String[] yearDay = lastCheckTime.split(":");
-        if (Integer.parseInt(yearDay[0]) >= year && Integer.parseInt(yearDay[1]) + 1 >= day) {
-          return;
-        }
-      }
-
-      ClientSetting.TRIPLEA_LAST_CHECK_FOR_ENGINE_UPDATE.save(year + ":" + day);
-      ClientSetting.flush();
-
-      final EngineVersionProperties latestEngineOut = EngineVersionProperties.contactServerForEngineVersionProperties();
-      if (latestEngineOut == null) {
-        return;
-      }
-      if (ClientContext.engineVersion().isLessThan(latestEngineOut.getLatestVersionOut())) {
-        SwingUtilities
-            .invokeLater(() -> EventThreadJOptionPane.showMessageDialog(null, latestEngineOut.getOutOfDateComponent(),
-                "Please Update TripleA", JOptionPane.INFORMATION_MESSAGE));
-      }
-    } catch (final Exception e) {
-      ClientLogger.logError("Error while checking for engine updates", e);
-    }
-  }
-
-  private static void checkForTutorialMap() {
-    final MapDownloadController mapDownloadController = ClientContext.mapDownloadController();
-    final boolean promptToDownloadTutorialMap = mapDownloadController.shouldPromptToDownloadTutorialMap();
-    mapDownloadController.preventPromptToDownloadTutorialMap();
-    if (!promptToDownloadTutorialMap) {
-      return;
-    }
-
-    final String message = "<html>Would you like to download the tutorial map?<br><br>"
-        + "(You can always download it later using the Download Maps<br>"
-        + "command if you don't want to do it now.)</html>";
-    SwingComponents.promptUser("Welcome to TripleA", message, () -> {
-      DownloadMapsWindow.showDownloadMapsWindowAndDownload("Tutorial");
-    });
   }
 }
